I think everything is working!
