+++ packages/base/src/index.ts
@@ -7,6 +7,7 @@ export * from './core/ISkeleton';
 export * from './core/TextureAtlas';
 export * from './core/TextureRegion';
 export * from './core/Utils';
+export * from './core/SkeletonBoundsBase';
 
 export * from './settings';
 export * from './SpineBase';
