+++ packages/runtime-3.7/src/core/attachments/RegionAttachment.ts
@@ -3,6 +3,7 @@ import {Attachment} from './Attachment';
 import {AttachmentType, ArrayLike, Color, TextureRegion, Utils, IRegionAttachment} from "@pixi-spine/base";
 
 import type {Bone} from '../Bone';
+import { Slot } from '../Slot';
 
 /**
  * @public
@@ -130,9 +131,9 @@ export class RegionAttachment extends Attachment implements IRegionAttachment {
         }
     }
 
-    computeWorldVertices(bone: Bone, worldVertices: ArrayLike<number>, offset: number, stride: number) {
+    computeWorldVertices(bone: Bone | Slot, worldVertices: ArrayLike<number>, offset: number, stride: number) {
         let vertexOffset = this.offset;
-        let mat = bone.matrix;
+        let mat = bone instanceof Slot? bone.bone.matrix : bone.matrix;
         let x = mat.tx, y = mat.ty;
         let a = mat.a, b = mat.c, c = mat.b, d = mat.d;
         let offsetX = 0, offsetY = 0;
